requirements-dev.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangotoolbox==1.8.0
dnspython==2.1.0
psycopg2-binary==2.9.1
pymongo==3.12.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
